Make server middleware standalone functions.
Removing direct dependencies from the server configuration.
